Update Hello.html
in this one Ive made the hello world bold. in order for the program to see my chages and register the bold command. you have to put it withing the 'text' variable in order to be seen. Ive out it in text bc ive made this section a variable but if itsnot a variable maybe put the bold command in the windoe alert ("") ??
ive put <b> and </b> on each end to open and close it