Add files via upload
Hi Simon its Seth I'm handing in my assignment.
I have special consideration for extra time I believe I talked to you about this a little before the quiz.
I'm sending it through email as the Canvas is closed. I do realize it is a little late even with my special consideration this was due to having to reset my computer.

I have provided my pseudo code in a word doc as well as the code for my portrait in word doc form.